everything compiles now without warnings on vs120 warning level 4

+++ krakenplay/implementation/inputmanager.cpp
@@ -74,7 +74,7 @@ namespace Krakenplay
 		HandleNewConnects(readState.gamepadStates, oldReadState.gamepadStates);
 	}
 
-	void InputManager::ReceiveInput(const MessageChunkHeader& header, const void* messageBody, unsigned int messageBodySize, unsigned int clientID)
+	void InputManager::ReceiveInput(const MessageChunkHeader& header, const void* messageBody, unsigned int messageBodySize, uint16_t clientID)
 	{
 		assert(messageBody != nullptr || (messageBody == nullptr && messageBodySize == 0) && "Invalid message body pointer!");
 		assert(GetMessageBodySize(header.messageType) == messageBodySize && "Invalid message body size!");
